requirements.txt
alembic==1.0.11
arrow==0.17.0
binaryornot==0.4.4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cookiecutter==1.7.2
dnspython==1.16.0
eventlet==0.28.0
fastapi==0.61.1
greenlet==0.4.17
h11==0.8.1
idna==2.10
Jinja2==2.11.2
jinja2-time==0.2.0
Mako==1.1.3
MarkupSafe==1.1.1
monotonic==1.5
poyo==0.5.0
psycopg2-binary==2.8.3
pydantic==1.6.1
python-dateutil==2.8.1
python-editor==1.0.4
python-engineio==3.13.2
python-slugify==4.0.1
python-socketio==4.6.0
PyYAML==5.1
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.6
starlette==0.13.6
text-unidecode==1.3
tornado==6.0.4
urllib3==1.25.10
uvicorn==0.8.4
websockets==7.0
